% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_integration.R
\name{align_scdna_scrna_genes}
\alias{align_scdna_scrna_genes}
\title{Align scDNA and scRNA matrices to common gene set}
\usage{
align_scdna_scrna_genes(
  scdna_gene_matrix,
  scrna_normalized_matrix,
  scdna_gene_subclone
)
}
\arguments{
\item{scdna_gene_matrix}{scDNA gene matrix}

\item{scrna_normalized_matrix}{Normalized scRNA matrix}

\item{scdna_gene_subclone}{scDNA clone matrix}
}
\value{
List of aligned matrices (scrna_normalized_matrix, scdna_gene_matrix, scdna_gene_subclone)
}
\description{
Align scDNA and scRNA matrices to common gene set
}
