% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alignment_score}
\alias{alignment_score}
\title{Computes alignment score}
\usage{
alignment_score(data, source_labels, target_labels, nn = 0)
}
\value{
Boolean
}
\description{
Computes alignment score
}
\keyword{internal}
