% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignSeuratWrapper.R
\name{as.SingleCellExperimentList}
\alias{as.SingleCellExperimentList}
\title{Converts a list of Seurat objects into a list of Single Cell Experiment objects
Extends the native Seurat function "as.SingleCellExperiment" by adding the capability to pass lists of Seurat objects}
\usage{
as.SingleCellExperimentList(seurat.obj, genes.use = NULL)
}
\arguments{
\item{seurat.obj}{List of Seurat objects to be converted to Single Cell Experiment objects.}

\item{genes.use}{A character vector of gene names which exhibit variable expression across all datasets}
}
\value{
A correctly initialized SingleCellExperiment object
}
\description{
Converts a list of Seurat objects into a list of Single Cell Experiment objects
Extends the native Seurat function "as.SingleCellExperiment" by adding the capability to pass lists of Seurat objects
}
\examples{

sce.objects = as.SingleCellExperimentList(seurat.obj)

}
\references{
Johansen, N., Quon, G. scAlign: a tool for alignment, integration, and rare cell identification from scRNA-seq data.
            Genome Biol 20, 166 (2019) doi: https://doi.org/10.1186/s13059-019-1766-4
}
\seealso{
<https://github.com/quon-titative-biology/scAlign>

\code{\link{ExtractGenes}}

\code{\link{SelectColors}}

\code{\link{DecoderVariance}}

\code{\link{MeanDecoderVariance}}

\code{\link{GetCharMetadata}}
}
\author{
Tyler Brassel, Nelson Johansen, Gerald Quon
}
