% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellbench}
\alias{cellbench}
\title{CellBench data for benchmarking}
\format{
An matrix object.
}
\source{
\href{https://github.com/LuyiTian/CellBench_data/tree/master/data}{CellBench}
}
\usage{
data(cellbench)
}
\description{
Data from RNA mixture experiments of CellBench which prove to be more
difficult alignment tasks.
}
\keyword{datasets}
