% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Center_cells_across_chromosome}
\alias{Center_cells_across_chromosome}
\title{Center cells across chromosome}
\usage{
Center_cells_across_chromosome(data, method = "median")
}
\arguments{
\item{data}{Matrix}

\item{method}{'median' or 'mean'}
}
\value{
Centered matrix
}
\description{
Center cells across chromosome
}
