% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_encoder.R, R/train_encoder_multi.R
\name{compute_kernel_matrix}
\alias{compute_kernel_matrix}
\title{Computes kernel matrix}
\usage{
compute_kernel_matrix(data, data_shape, labels = NULL, method = NULL,
  perplexity = 30, diag = "zero")

compute_kernel_matrix(data, data_shape, labels = NULL, method = NULL,
  perplexity = 30, diag = "zero")
}
\arguments{
\item{data}{cell x feature data matrix}

\item{data_shape}{number of features for data}

\item{labels}{cell x 1 annotation (label) vector}

\item{method}{Kernel to compute pairwise cell similarity}

\item{perplexity}{neighborhood parameter for gaussian kernel}

\item{diag}{indicator for self similarity}

\item{data}{cell x feature data matrix}

\item{data_shape}{number of features for data}

\item{labels}{cell x 1 annotation (label) vector}

\item{method}{Kernel to compute pairwise cell similarity}

\item{perplexity}{neighborhood parameter for gaussian kernel}

\item{diag}{indicator for self similarity}
}
\value{
Tensorflow op

Tensorflow op
}
\description{
Adds the defined kernel to the operation graph for high dimensional pairwise data similarity

Adds the defined kernel to the operation graph for high dimensional pairwise data similarity
}
\keyword{internal}
