% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architectures.R
\name{decoder_medium}
\alias{decoder_medium}
\title{Network medium}
\usage{
decoder_medium(inputs, complexity, final_dim, emb_size = 32,
  l2_weight = 1e-04, dropout_rate = 0.3, is_training = TRUE,
  batch_norm = FALSE)
}
\value{
Neural network graph op
}
\description{
Defines network architecture for scAlign.
}
\keyword{internal}
