% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_decoder.R
\name{decoderModel_add_mse_loss}
\alias{decoderModel_add_mse_loss}
\title{Convert embedding to logit scores}
\usage{
decoderModel_add_mse_loss(proj, data, mode, debug = TRUE)
}
\arguments{
\item{proj}{Current project of data during training}

\item{data}{High dimensional data}

\item{mode}{Dataste being operated on}

\item{debug}{Verbosity of command line output}
}
\value{
Tensorflow op
}
\description{
Convert embedding to logit scores
}
\keyword{internal}
