% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_decoder.R
\name{decoderModel_calc_projection}
\alias{decoderModel_calc_projection}
\title{Compute projections for complete data after training}
\usage{
decoderModel_calc_projection(sess, data, data_full, endpoint, test_in,
  FLAGS, batch_size = 150)
}
\arguments{
\item{sess}{Current tensorflow session}

\item{data}{Input matrix or tensor to reduce to embedding space}

\item{endpoint}{Operation to evaluate in the graph}

\item{test_in}{Placeholder to process full data in batches}

\item{FLAGS}{Tensorflow run arguments}

\item{batch_size}{how many cells per each batch}
}
\value{
Tensorflow op
}
\description{
Takes the embeddings and returns high dimensional data
}
\keyword{internal}
