% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_decoder.R
\name{decoderModel_emb_to_proj}
\alias{decoderModel_emb_to_proj}
\title{Compute projection for mini-batch}
\usage{
decoderModel_emb_to_proj(model_func, data, is_training = TRUE)
}
\arguments{
\item{model_func}{Network op in graph}

\item{data}{Data to pass through model_func}

\item{is_training}{Determines whether dropout and batch norm are run.}
}
\value{
Tensorflow op
}
\description{
Takes the embeddings and returns high dimensional data
}
\keyword{internal}
