% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_decoder.R
\name{decoderModel_train_decoder}
\alias{decoderModel_train_decoder}
\title{decoder for scAlign}
\usage{
decoderModel_train_decoder(
  FLAGS,
  config,
  mode,
  data_full,
  data_emb,
  all_data_emb
)
}
\arguments{
\item{FLAGS}{Tensorflow run arguments.}

\item{config}{Hardware configuration options.}

\item{data_full}{High dimensional data used for reconstruction error}

\item{data_emb}{Low dimensional data from scAlign encoder post alignment}

\item{all_data_emb}{Low dimensional cells from all conditions to project}

\item{CV}{Model that is being run.}
}
\value{
final interpolation results
}
\description{
Projects cells from joint cell state space into one condition.
}
\keyword{internal}
