% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignSeuratWrapper.R
\name{DecoderVariance}
\alias{DecoderVariance}
\title{Creates a 2d array of variance values calculated across the 3rd decoder dimension of a 3d array of decoder matrices.
Unless specified, automatically calculates highly variable genes and orders the combined variance matrix in decreasing order,
for later graphical analysis analysis.}
\usage{
DecoderVariance(
  all_data_combined,
  seurat.obj,
  genes.use,
  output.length,
  decreasing = TRUE,
  return.genes = TRUE
)
}
\arguments{
\item{all_data_combined}{A 3D matrix of decoder values across conditions, (cells x genes x condition)}

\item{seurat.obj}{the reference, aligned Seurat object used to correctly populate the rownames of the output array}

\item{genes.use}{The gene names to use. See ExtractGenes for more information.}

\item{output.length}{The number of genes to output, important for later graphical analysis.}

\item{decreasing}{A logical value for ordering the 2nd value of the output list. Setting to TRUE will output genes in order from highest to lowest mean variance.}

\item{return.genes}{A logical vector defaulted to TRUE. Setting to FALSE will only output the all_data_var, not high_var_genes}
}
\value{
A list of arrays, the fist being the variance calculated across decoders, the second a list of order gene indices
}
\description{
Creates a 2d array of variance values calculated across the 3rd decoder dimension of a 3d array of decoder matrices.
Unless specified, automatically calculates highly variable genes and orders the combined variance matrix in decreasing order,
for later graphical analysis analysis.
}
\examples{
        
}
\seealso{
\code{\link{as.SingleCellExperimentList}}

\code{\link{ExtractGenes}}

\code{\link{SelectColors}}

\code{\link{MeanDecoderVariance}}

\code{\link{GetCharMetadata}}
}
