% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_encoder.R
\name{define_kernel_matrix}
\alias{define_kernel_matrix}
\title{define kernel matrix}
\usage{
define_kernel_matrix(
  method,
  data = NULL,
  labels = NULL,
  data_shape,
  perplexity = NULL,
  diag = "zero"
)
}
\arguments{
\item{method}{Kernel to compute pairwise cell similarity}

\item{data}{cell x feature data matrix}

\item{labels}{cell x 1 annotation (label) vector}

\item{data_shape}{number of features for data}

\item{perplexity}{neighborhood parameter for gaussian kernel}

\item{diag}{indicator for self similarity}
}
\value{
Tensorflow op
}
\description{
Adds the defined kernel to the operation graph for high dimensional pairwise data similarity
}
\keyword{internal}
