% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.early_stop}
\alias{.early_stop}
\title{Perform early stopping}
\usage{
.early_stop(loss_tracker, step, patience_count, early_stopping_active,
  patience = 30, min_delta = 0.01)
}
\value{
Earlying stopping flag
}
\description{
Perform early stopping
}
\keyword{internal}
