% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is.wholenumber}
\alias{.is.wholenumber}
\title{Check for whole number}
\usage{
.is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\value{
Boolean
}
\description{
Check for whole number
}
\keyword{internal}
