% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.learning_rate}
\alias{.learning_rate}
\title{Sets the learning rate for optimizer}
\usage{
.learning_rate(step, decay_step, FLAGS)
}
\value{
Tensorflow op
}
\description{
Sets the learning rate for optimizer
}
\keyword{internal}
