% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.plotTSNE}
\alias{.plotTSNE}
\title{Creates tsne plot}
\usage{
.plotTSNE(
  data,
  labels,
  file_out = file.path(tempdir(), "/scAlign_default_plot.png")
)
}
\arguments{
\item{data}{Current aligned embeddings}

\item{labels}{Object labels}

\item{file_out}{File name to save plot}
}
\value{
ggplot2 object
}
\description{
Creates tsne plot
}
\keyword{internal}
