% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_add_logit_loss}
\alias{encoderModel_add_logit_loss}
\title{Add classifier loss component}
\usage{
encoderModel_add_logit_loss(logits, labels, weight = 1, smoothing = 0)
}
\arguments{
\item{logits}{cell x label logits}

\item{labels}{cell x 1 true labels}

\item{weight}{Scale of classifier loss component}

\item{smoothing}{tensorflow label_smoothing (unused)}
}
\value{
Tensorflow loss op
}
\description{
Cross entropy of labels and predictions
}
\keyword{internal}
