% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_add_semisup_loss_data_driven}
\alias{encoderModel_add_semisup_loss_data_driven}
\title{Add walker loss component}
\usage{
encoderModel_add_semisup_loss_data_driven(p_target, a, b, data_shape,
  walker_weight = 1, visit_weight = 1, mode = "source",
  comp = "exp", betas = NULL, diag = "zero", debug = TRUE)
}
\arguments{
\item{p_target}{Pairwise cell-cell similarity}

\item{a}{Data embeddings}

\item{b}{Data embeddings}

\item{data_shape}{Dimension of data}

\item{walker_weight}{Walker loss component weight}

\item{visit_weight}{Visit loss component weight (unused)}

\item{mode}{Specifies which walker loss is being computed}

\item{comp}{How to compute round trip probabilities}

\item{betas}{Value for fixed beta gaussian kernel}

\item{diag}{Whether to include self similarity}

\item{debug}{Sets the verbosity level}
}
\value{
Tensorflow loss op
}
\description{
Creates the walker loss component using embedding representation of
high-dimensional data. Specifies random walk procedure.
}
\keyword{internal}
