% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_calc_embedding}
\alias{encoderModel_calc_embedding}
\title{Compute embedding for complete data after training}
\usage{
encoderModel_calc_embedding(
  sess,
  data,
  endpoint,
  test_in,
  FLAGS,
  batch_size = 150
)
}
\arguments{
\item{sess}{Current tensorflow session}

\item{data}{Input matrix or tensor to reduce to embedding space}

\item{endpoint}{Operation to evaluate in the graph}

\item{test_in}{Placeholder to process full data in batches}

\item{FLAGS}{Tensorflow run arguments}

\item{batch_size}{how many cells per each batch}
}
\value{
Aligned data embedding
}
\description{
Takes the high dimensional data through the neural network to embedding layer
}
\keyword{internal}
