% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_create_train_op}
\alias{encoderModel_create_train_op}
\title{Builds training operation}
\usage{
encoderModel_create_train_op(learning_rate, step)
}
\arguments{
\item{learning_rate}{Learning rate for optimizer}

\item{step}{Global training step}
}
\value{
Tensorflow training op
}
\description{
Defines how the model is optimized and update scheme.
}
\keyword{internal}
