% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_data_to_embedding}
\alias{encoderModel_data_to_embedding}
\title{Compute embedding for mini-batch}
\usage{
encoderModel_data_to_embedding(model_func, data, is_training = TRUE,
  scope = "")
}
\arguments{
\item{model_func}{Network op in graph}

\item{data}{Data to pass through model_func}

\item{is_training}{Determines whether dropout and batch norm are run.}
}
\value{
Tensorflow op
}
\description{
Takes the high dimensional data through the neural network to embedding layer
}
\keyword{internal}
