% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_embedding_to_logit}
\alias{encoderModel_embedding_to_logit}
\title{Convert embedding to logit scores}
\usage{
encoderModel_embedding_to_logit(data, num_labels, is_training = TRUE)
}
\arguments{
\item{data}{Data to pass through model_func.}

\item{num_labels}{How many unique labels there are.}

\item{is_training}{Determines whether dropout and batch norm are run.}
}
\value{
Tensorflow op
}
\description{
Convert embedding to logit scores
}
\keyword{internal}
