% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_encoder.R
\name{encoderModel_gaussian_kernel}
\alias{encoderModel_gaussian_kernel}
\title{Guassian kernel}
\usage{
encoderModel_gaussian_kernel(data, data_b = NULL, dim, perplexity = 30,
  method = "euclidean", betas.fixed = 0, diag = "zero")
}
\arguments{
\item{data}{cell x feature data matrix}

\item{data_b}{second cell x feature data matrix, if none provided then data_b = data}

\item{dim}{number of cells in of mini-batch}

\item{perplexity}{neighborhood parameter for gaussian kernel}

\item{method}{distance metric prior to computing probabilities}

\item{diag}{indicator for self similarity}
}
\value{
Tensorflow op
}
\description{
Tensorflow implementation of tsne's gaussian_kernel, accepts tensors of
float64 (for precision) and returns tensors of float32
}
\keyword{internal}
