% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_encoder_multi.R
\name{encoderModel_train_encoder_multi}
\alias{encoderModel_train_encoder_multi}
\title{encoder for scAlign}
\usage{
encoderModel_train_encoder_multi(
  FLAGS,
  CV,
  config,
  num_labels,
  data_shape,
  names,
  data,
  labels
)
}
\arguments{
\item{FLAGS}{Tensorflow run arguments.}

\item{CV}{Model that is being run.}

\item{config}{Hardware configuration options.}

\item{num_labels}{Number of unique labels.}

\item{data_shape}{Dimensions of input data.}

\item{source_data}{Source data matrix (cell x feature).}

\item{source_labels}{Source data labels.}

\item{target_data}{Target data matrix (cell x feature).}

\item{target_labels}{Target data labels.}
}
\value{
Final alignment results
}
\description{
Aligns paired data into a joint embedding space.
}
\keyword{internal}
