% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_seurat_processing.R
\name{extract_epithelial_cells}
\alias{extract_epithelial_cells}
\title{Extract epithelial cells and rerun Seurat pipeline}
\usage{
extract_epithelial_cells(seu_merged, ndims = 100)
}
\arguments{
\item{seu_merged}{Processed Seurat object}

\item{ndims}{Number of dimensions for PCA/UMAP (default 100)}
}
\value{
Epithelial Seurat object
}
\description{
Extract epithelial cells and rerun Seurat pipeline
}
