% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignSeuratWrapper.R
\name{ExtractGenes}
\alias{ExtractGenes}
\title{Extracts common set of genes across all datasets}
\usage{
ExtractGenes(seurat.obj1, seurat.obj2 = NULL)
}
\arguments{
\item{seurat.obj1}{Either a preprocessed Seurat object or a list of Seurat objects}

\item{seurat.obj2}{A preprocessed Seurat object (optional, for simple binary conditions)}
}
\value{
A character vector of gene names which exhibit varible expression across all datasets
}
\description{
Extracts common set of genes across all datasets
}
\examples{
for (i in 1:length(seurat.list)) {
  seurat.list[[i]] <- NormalizeData(seurat.list[[i]])
  seurat.list[[i]] <- ScaleData(seurat.list[[i]], do.scale=T, do.center=T, display.progress=T)
  seurat.list[[i]] <- FindVariableFeatures(seurat.list[[i]], nFeatures = 3000)
}
genes.use <- ExtractGenes(seurat.list)

}
\seealso{
\code{\link{as.SingleCellExperimentList}}

\code{\link{SelectColors}}

\code{\link{DecoderVariance}}

\code{\link{MeanDecoderVariance}}

\code{\link{GetCharMetadata}}
}
