% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Filter_genes}
\alias{Filter_genes}
\title{Filter genes expressed in fewer than min_num_cells}
\usage{
Filter_genes(mat, min_num_cells = 3)
}
\arguments{
\item{mat}{Expression matrix}

\item{min_num_cells}{Integer, minimum number of cells}
}
\value{
Filtered matrix
}
\description{
Filter genes expressed in fewer than min_num_cells
}
