% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_seurat_processing.R
\name{find_and_plot_markers}
\alias{find_and_plot_markers}
\title{Find markers and plot heatmaps for merged Seurat object}
\usage{
find_and_plot_markers(aggr, output_dir = ".")
}
\arguments{
\item{aggr}{Processed Seurat object}

\item{output_dir}{Directory to save output files, default is current working directory}
}
\description{
Find markers and plot heatmaps for merged Seurat object
}
