% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Computes gaussian kernel matrix}
\usage{
gaussianKernel(data, data_shape, labels = NULL, method = NULL,
  perplexity = 30, diag = "zero")
}
\arguments{
\item{data}{cell x feature data matrix}

\item{data_shape}{number of features for data}

\item{labels}{cell x 1 annotation (label) vector}

\item{method}{Kernel to compute pairwise cell similarity}

\item{perplexity}{neighborhood parameter for gaussian kernel}

\item{diag}{indicator for self similarity}
}
\value{
Tensorflow op
}
\description{
Tensorflow implementation of tSNE's gaussian kernel.
}
\examples{

## Input data, 100 cells x 10 features
data = matrix(sample.int(1000, 100*10, TRUE), 100, 10)
rownames(data) = paste0("cell", seq_len(100))
colnames(data) = paste0("gene", seq_len(10))

result = gaussianKernel(data, nrow(data))

}
