% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignSeuratWrapper.R
\name{GetCharMetadata}
\alias{GetCharMetadata}
\title{Fetch metadata as a character vector}
\usage{
GetCharMetadata(seurat.obj, value = "celltype", unique = TRUE)
}
\arguments{
\item{seurat.obj}{A seurat object}

\item{value}{The name of the internal metadata object the function should access}
}
\value{
A character vector of the unique values accessed by the Seurat function FetchData
}
\description{
Fetch metadata as a character vector
}
\seealso{
\code{\link{as.SingleCellExperimentList}}

\code{\link{ExtractGenes}}

\code{\link{SelectColors}}

\code{\link{DecoderVariance}}

\code{\link{MeanDecoderVariance}}
}
