% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCCA.R
\name{GetCrit}
\alias{GetCrit}
\title{Seurat V2 MultiCCA}
\usage{
GetCrit(mat.list, ws, num.sets)
}
\arguments{
\item{mat.list}{List of matrix objects}

\item{ws}{optim}

\item{num.sets}{Number of sets}
}
\value{
crit values
}
\description{
Seurat V2 MultiCCA
}
\keyword{internal}
