% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdna_components.R
\name{identify_cellranger_noise}
\alias{identify_cellranger_noise}
\title{Identify cellranger noise cells based on cellranger-dna metrics}
\usage{
identify_cellranger_noise(
  per_cell_metrics_file = "per_cell_summary_metrics.csv",
  seurat_obj = "seurat_scDNA.rds"
)
}
\arguments{
\item{per_cell_metrics_file}{Path to per_cell_summary_metrics.csv file}

\item{seurat_obj}{Seurat object or path to seurat_scDNA.rds file}
}
\value{
Seurat object with celltype annotation
}
\description{
Identify cellranger noise cells based on cellranger-dna metrics
}
