% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdna_components.R
\name{identify_normal_cells}
\alias{identify_normal_cells}
\title{Identify normal cells based on chromosome-wise CNV analysis}
\usage{
identify_normal_cells(
  seurat_obj = "seurat_scDNA.rds",
  scdna_matrix_locs,
  cnv_threshold = 2.5
)
}
\arguments{
\item{seurat_obj}{Seurat object with celltype annotation}

\item{scdna_matrix_locs}{Data frame with chromosome information for segments}

\item{cnv_threshold}{Threshold for CNV value to identify tumor cells, default 2.5}
}
\value{
Seurat object with updated celltype annotation
}
\description{
Identify normal cells based on chromosome-wise CNV analysis
}
