% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Invert_logtransform}
\alias{Invert_logtransform}
\title{Invert log2 transform}
\usage{
Invert_logtransform(data)
}
\arguments{
\item{data}{Matrix}
}
\value{
Inverse log2 transformed matrix
}
\description{
Invert log2 transform
}
