% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Logtransform}
\alias{Logtransform}
\title{Log2 transform}
\usage{
Logtransform(data)
}
\arguments{
\item{data}{Matrix}
}
\value{
Log2 transformed matrix
}
\description{
Log2 transform
}
