% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Max_threshold_bounds}
\alias{Max_threshold_bounds}
\title{Apply max/min threshold bounds}
\usage{
Max_threshold_bounds(data, threshold = 3)
}
\arguments{
\item{data}{Matrix}

\item{threshold}{Numeric}
}
\value{
Thresholded matrix
}
\description{
Apply max/min threshold bounds
}
