% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignSeuratWrapper.R
\name{MeanDecoderVariance}
\alias{MeanDecoderVariance}
\title{Calculate mean decoder variance}
\usage{
MeanDecoderVariance(all_data_var)
}
\arguments{
\item{all_data_var}{A matrix of variance values calculated across the decoder dimension.}
}
\value{
A vector of mean data values for each column of the "all_data_var" parameter.
}
\description{
Calculate mean decoder variance
}
\seealso{
\code{\link{as.SingleCellExperimentList}}

\code{\link{ExtractGenes}}

\code{\link{SelectColors}}

\code{\link{DecoderVariance}}

\code{\link{GetCharMetadata}}
}
