% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{normalize_counts_by_seq_depth}
\alias{normalize_counts_by_seq_depth}
\title{Normalize counts by sequencing depth}
\usage{
normalize_counts_by_seq_depth(mat, normalize_factor = NA)
}
\arguments{
\item{mat}{Expression matrix}

\item{normalize_factor}{Numeric, normalization factor (default: median of column sums)}
}
\value{
Normalized matrix
}
\description{
Normalize counts by sequencing depth
}
