% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Normalize_RNA}
\alias{Normalize_RNA}
\title{Full RNA normalization pipeline}
\usage{
Normalize_RNA(scRNA_raw_matrix, gene_locs)
}
\arguments{
\item{scRNA_raw_matrix}{Raw expression matrix}

\item{gene_locs}{Data.frame with gene locations}
}
\value{
Normalized matrix
}
\description{
Full RNA normalization pipeline
}
