% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_seurat_processing.R
\name{process_merged_seurat}
\alias{process_merged_seurat}
\title{SCTransform, PCA, clustering, UMAP for merged Seurat object}
\usage{
process_merged_seurat(seu_merged, dims.reduce = 20, cluster.res = 1)
}
\arguments{
\item{seu_merged}{Merged Seurat object}

\item{dims.reduce}{Number of dimensions for reduction (default 20)}

\item{cluster.res}{Clustering resolution (default 1)}
}
\value{
Processed Seurat object
}
\description{
SCTransform, PCA, clustering, UMAP for merged Seurat object
}
