% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_integration.R
\name{project_scrna_to_scdna}
\alias{project_scrna_to_scdna}
\title{Project scRNA expression into scDNA space using linear mapping}
\usage{
project_scrna_to_scdna(scrna_normalized_matrix, scdna_gene_matrix)
}
\arguments{
\item{scrna_normalized_matrix}{scRNA expression matrix (genes x cells)}

\item{scdna_gene_matrix}{scDNA CNV matrix (genes x cells)}
}
\value{
Projected scRNA matrix in scDNA space
}
\description{
Project scRNA expression into scDNA space using linear mapping
}
