% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdna_components.R
\name{reidentify_replicates}
\alias{reidentify_replicates}
\title{Re-identify replicates from cellranger noise and S phase cells}
\usage{
reidentify_replicates(
  seurat_obj = "seurat_scDNA.rds",
  mean_threshold = 2.5,
  auto_threshold = "mannual"
)
}
\arguments{
\item{seurat_obj}{Seurat object with celltype annotation, or path to RDS file}

\item{mean_threshold}{Threshold for mean CNV value to distinguish S phase from noise, default 2.5}

\item{auto_threshold}{Whether to automatically determine threshold from data, default 'manual'}
}
\value{
Seurat object with updated celltype annotation
}
\description{
Re-identify replicates from cellranger noise and S phase cells
}
