% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{run_scrna_normalization}
\alias{run_scrna_normalization}
\title{Main scRNA normalization pipeline}
\usage{
run_scrna_normalization(
  seurat_epi,
  gene_locs_path,
  output_rds = "RNA_normalized_matrix.rds",
  output_dir = "."
)
}
\arguments{
\item{seurat_epi}{Seurat RDS file}

\item{gene_locs_path}{Path to gene location file}

\item{output_rds}{Path to save normalized matrix}

\item{output_dir}{Directory to save output files, default is current working directory}
}
\value{
Normalized matrix (invisible)
}
\description{
Main scRNA normalization pipeline
}
