% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCCA.R
\name{RunMultiCCA}
\alias{RunMultiCCA}
\title{Seurat V2 MultiCCA}
\usage{
RunMultiCCA(mat.list, niter = 25, num.ccs = 10, standardize = TRUE)
}
\arguments{
\item{mat.list}{List of matrix objects}

\item{niter}{Optimization parameter}

\item{num.ccs}{Number of dimensions}

\item{standardize}{To standardize data}
}
\value{
CCA results
}
\description{
Seurat V2 MultiCCA
}
\keyword{internal}
