% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign.R
\name{scAlign}
\alias{scAlign}
\title{Run scAlign framework}
\usage{
scAlign(object, options = scAlignOptions(),
  encoder.data = "scale.data", decoder.data = encoder.data,
  supervised = "none", run.encoder = TRUE, run.decoder = TRUE,
  log.dir = "./models/", log.results = FALSE, device = "CPU")
}
\arguments{
\item{object}{scAlign object.}

\item{options}{Training options for scAlign.}

\item{encoder.data}{Which data format to use for alignment.}

\item{decoder.data}{Which data format to use for interpolation.}

\item{supervised}{Run scAlign in supervised mode, requires labels.}

\item{run.encoder}{Run scAlign alignment procedure.}

\item{run.decoder}{Run scAlign projection through paired decoders.}

\item{log.dir}{Location to save results.}

\item{log.results}{Determines if results should be written to log.dir.}

\item{device}{Specify hardware to use.}
}
\value{
scAlign class
}
\description{
Main function for scAlign that runs encoder and decoder networks
}
\examples{

library(Seurat)

## Input data, 1000 genes x 100 cells
data = matrix(sample.int(10000, 1000*100, TRUE), 1000, 100)
rownames(data) = paste0("gene", seq_len(1000))
colnames(data) = paste0("cell", seq_len(100))

age    = c(rep("young",50), rep("old",50))
labels = c(c(rep("type1",25), rep("type2",25)), c(rep("type1",25), rep("type2",25)))

ctrl.data = data[,which(age == "young")]
stim.data = data[,which(age == "old")]

ctrl = CreateSeuratObject(raw.data = ctrl.data, project = "MOUSE_AGE", min.cells = 0)
ctrl@meta.data$stim  = "YOUNG"
ctrl@meta.data$label = labels[which(age == "young")]
ctrl = ScaleData(ctrl, do.scale=FALSE, do.center=FALSE, scale.max=50, display.progress = TRUE)

stim = CreateSeuratObject(raw.data = stim.data, project = "MOUSE_AGE", min.cells = 0)
stim@meta.data$stim = "OLD"
stim@meta.data$label = labels[which(age == "old")]
stim = ScaleData(stim, do.scale=FALSE, do.center=FALSE, scale.max=50, display.progress = TRUE)

## Build the scAlign class object and compute PCs
scAlignHSC = scAlignCreateObject(objects = list("YOUNG"=ctrl, "OLD"=stim),
                                 labels = list(ctrl@meta.data$label, stim@meta.data$label),
                                 pca.reduce = FALSE,
                                 pcs.compute = 50,
                                 cca.reduce = FALSE,
                                 ccs.compute = 15,
                                 project.name = "scAlign_example")

## Run scAlign with high_var_genes
scAlignHSC = scAlign(scAlignHSC,
                     options=scAlignOptions(steps=1, log.every=1, early.stop=FALSE, architecture="large"),
                     encoder.data="scale.data",
                     supervised='none',
                     run.encoder=TRUE,
                     run.decoder=FALSE,
                     log.results=FALSE,
                     device="CPU")

}
