% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{scAlignArguments}
\alias{scAlignArguments}
\title{Record aguments passed to scAlign}
\usage{
scAlignArguments(sce.object, encoder.data, decoder.data, supervised,
  run.encoder, run.decoder, log.dir, log.results, device)
}
\value{
Arguments data.frame
}
\description{
Record aguments passed to scAlign
}
\keyword{internal}
