% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{class}
\name{scAlignClass-class}
\alias{scAlignClass-class}
\alias{scAlignClass}
\title{scAlign S4 classs}
\value{
scAlign object
}
\description{
Holds input data and resulting data from scAlign.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{List containing unaligned (raw) datasets.}

\item{\code{scale.data}}{List containing unaligned (raw) datasets.}

\item{\code{reduced.data}}{List containing input data after an intial step of dimensionality reduction.}

\item{\code{aligned.data}}{List containing aligned datasets.}

\item{\code{projected.data}}{List containing the set of projections for each dataset.}

\item{\code{genes.use}}{Vector of genes used to subset input data.}

\item{\code{meta.data}}{data.frame containing dataset identifiers and cell information.}

\item{\code{options}}{data.frame of parameters essential to neural network training.}

\item{\code{arguments}}{data.frame recording arguments for scAlign function.}

\item{\code{project.name}}{Name for the scAlign project.}

\item{\code{version}}{Current version of scAlign.}
}}

