% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{scAlignCreateObject}
\alias{scAlignCreateObject}
\title{Creates scAlign object}
\usage{
scAlignCreateObject(sce.objects, labels = list(), meta.data = NULL,
  pca.reduce = TRUE, pcs.compute = 20, cca.reduce = TRUE,
  ccs.compute = 15, data.use = "scale.data",
  project.name = "scAlignProject")
}
\arguments{
\item{sce.objects}{List of Seurat or Matrix objects; sample x feature.}

\item{labels}{List of labels for each object.}

\item{meta.data}{Additional meta.data to add.}

\item{pca.reduce}{Initial step of dimensionality be performced by PCA.}

\item{pcs.compute}{Number of PCs to retrain for alignment.}

\item{cca.reduce}{Initial step of dimensionality be performced by CCA.}

\item{ccs.compute}{Number of CCs to retrain for alignment.}

\item{data.use}{Specificies which data to use from a Seurat object for dimensionality reduction.}

\item{project.name}{Name for current scAlign project.}
}
\value{
Initialized scAlign object
}
\description{
Creates scAlign object
}
\examples{

library(Seurat)
library(SingleCellExperiment)

## Input data, 1000 genes x 100 cells
data = matrix(sample.int(10000, 1000*100, TRUE), 1000, 100)
rownames(data) = paste0("gene", seq_len(1000))
colnames(data) = paste0("cell", seq_len(100))

age    = c(rep("young",50), rep("old",50))
labels = c(c(rep("type1",25), rep("type2",25)), c(rep("type1",25), rep("type2",25)))

ctrl.data = data[,which(age == "young")]
stim.data = data[,which(age == "old")]

ctrl = CreateSeuratObject(raw.data = ctrl.data, project = "MOUSE_AGE", min.cells = 0)
ctrl = ScaleData(ctrl, do.scale=TRUE, do.center=TRUE, scale.max=50, display.progress = TRUE)

stim = CreateSeuratObject(raw.data = stim.data, project = "MOUSE_AGE", min.cells = 0)
stim = ScaleData(stim, do.scale=TRUE, do.center=TRUE, scale.max=50, display.progress = TRUE)

## Build the SCE object for input to scAlign using Seurat preprocessing and variable gene selection
ctrlSCE <- SingleCellExperiment(
              assays = list(
                counts = t(FetchData(ctrl, vars.all=rownames(data), use.raw=TRUE)),
                scale.data = t(FetchData(ctrl, vars.all=rownames(data), use.scaled=TRUE))))

stimSCE <- SingleCellExperiment(
              assays = list(
                counts = t(FetchData(stim, vars.all=rownames(data), use.raw=TRUE)),
                scale.data = t(FetchData(stim, vars.all=rownames(data), use.scaled=TRUE))))

## Build the scAlign class object and compute PCs
scAlignHSC = scAlignCreateObject(sce.objects = list("YOUNG"=ctrlSCE,
                                                    "OLD"=stimSCE),
                                 labels = list(labels[which(age == "young")],
                                               labels[which(age == "old")]),
                                 pca.reduce = TRUE,
                                 pcs.compute = 50,
                                 cca.reduce = TRUE,
                                 ccs.compute = 15,
                                 project.name = "scAlign_Kowalcyzk_HSC")

}
