% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{class}
\name{scAlignEmbedding-class}
\alias{scAlignEmbedding-class}
\alias{scAlignEmbedding}
\title{Object to store embeddings}
\value{
scAlignEmbedding object
}
\description{
Object to store embeddings
}
\keyword{internal}
