% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{scAlignEmbeddingCreateObject}
\alias{scAlignEmbeddingCreateObject}
\title{Initialize object of class}
\usage{
scAlignEmbeddingCreateObject(embedding, method.results, type)
}
\value{
Initialized scAlignEmbedding object
}
\description{
Initialize object of class
}
\keyword{internal}
