% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{scAlignOptions}
\alias{scAlignOptions}
\title{Set training options}
\usage{
scAlignOptions(steps = 15000, batch.size = 150,
  learning.rate = 1e-04, log.every = 5000, architecture = "small",
  num.dim = 32, perplexity = 30, norm = TRUE, full_norm = FALSE,
  early.stop = TRUE, seed = 1234)
}
\arguments{
\item{steps}{(default: 15000) Number of training iterations for neural networks.}

\item{batch.size}{(default: 150) Number of input samples per training batch.}

\item{learning.rate}{(default: 1e-4) Initial learning rate for ADAM.}

\item{log.every}{(default: 5000) Number of steps before saving results.}

\item{architecture}{(default: "small") Network function name for scAlign.}

\item{num.dim}{(default: 32) Number of dimensions for joint embedding space.}

\item{perplexity}{(default: 30) Determines the neighborhood size for each sample.}

\item{norm}{(default: TRUE) Normalize the data mini batches while training scAlign (repeated).}

\item{full_norm}{(default: FALSE) Normalize the data matrix prior to scAlign (done once).}

\item{early.stop}{(default: TRUE) Early stopping during network training.}

\item{seed}{(default: 1245) Sets graph level random seed in tensorflow.}
}
\value{
Options data.frame
}
\description{
Defines parameters for optimizer and training procedure.
}
\examples{

options=scAlignOptions(steps=15000,
                       log.every=5000,
                       early.stop=FALSE,
                       architecture="large")

}
