% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{scAlignSetEmbedding}
\alias{scAlignSetEmbedding}
\title{Set a new embedding}
\usage{
scAlignSetEmbedding(object, emb, emb.name, emb.results = NULL)
}
\arguments{
\item{object}{scAlign S4 class object}

\item{emb}{Matrix of cells x features}

\item{emb.name}{Name of dimensionality reduction}

\item{emb.results}{Other results from dimensionality reduction method.}
}
\value{
scAlign object with embedding slot set

Adds additional dimensionality reduced data to scAlign class
}
\description{
Set a new embedding
}
\keyword{internal}
