% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignSeuratWrapper.R
\name{SelectColors}
\alias{SelectColors}
\title{Creates a list of unique color values used for plotting}
\usage{
SelectColors(seurat.obj, gradient = FALSE, value = "celltype")
}
\arguments{
\item{seurat.obj}{A singular preprocessed Seurat object}

\item{gradient}{Setting to TRUE will use a sequence of hsv colors instead of 20 unique colors,
useful for comparisons of more than 20 cell types.}

\item{value}{The Seurat metadata slot to generate colors for. Defaults to "celltype".}
}
\value{
A named vector of unique hexedecimal color values, either generated from a preselected
        vector of 20 unique colors, or from a sequence of colors in hsv colorspace.
}
\description{
Creates a list of unique color values used for plotting
}
\examples{
DimPlot(object = seurat.obj,
        reduction = "tsne",
        cols = SelectColors(seurat.obj),
        group.by = "celltype",
        label = TRUE,
        repel = TRUE)

}
\seealso{
\code{\link{as.SingleCellExperimentList}}

\code{\link{ExtractGenes}}

\code{\link{DecoderVariance}}

\code{\link{MeanDecoderVariance}}

\code{\link{GetCharMetadata}}
}
