% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{methods}
\name{setAligned,scAlignClass-method}
\alias{setAligned,scAlignClass-method}
\title{Method assigns aligned data to S4 slot}
\usage{
\S4method{setAligned}{scAlignClass}(self, aligned.data, data.use)
}
\value{
scAlign class with alignment slot set
}
\description{
Method assigns aligned data to S4 slot
}
\keyword{internal}
