% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{setAligned}
\alias{setAligned}
\title{Generic assigns aligned data to S4 slot}
\usage{
setAligned(self, aligned.data, data.use)
}
\value{
scAlign class with alignment slot set
}
\description{
Generic assigns aligned data to S4 slot
}
\keyword{internal}
