% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{methods}
\name{setArgs,scAlignClass-method}
\alias{setArgs,scAlignClass-method}
\title{Method assigns projected data to S4 slots}
\usage{
\S4method{setArgs}{scAlignClass}(self, args)
}
\value{
scAlign class with arguments slot set
}
\description{
Method assigns projected data to S4 slots
}
\keyword{internal}
