% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{setArgs}
\alias{setArgs}
\title{Generic assigns projected data to S4 slots}
\usage{
setArgs(self, args)
}
\value{
scAlign class with arguments slot set
}
\description{
Generic assigns projected data to S4 slots
}
\keyword{internal}
