% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{methods}
\name{setEmbedding,scAlignClass-method}
\alias{setEmbedding,scAlignClass-method}
\title{Method assigns aligned data to S4 slot}
\usage{
\S4method{setEmbedding}{scAlignClass}(self, emb, emb.results, emb.name)
}
\value{
scAlign class with reduced.data slot set
}
\description{
Method assigns aligned data to S4 slot
}
\keyword{internal}
