% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{setEmbedding}
\alias{setEmbedding}
\title{Generic assigns aligned data to S4 slot}
\usage{
setEmbedding(self, emb, emb.results, emb.name)
}
\value{
scAlign class with reduced.data slot set
}
\description{
Generic assigns aligned data to S4 slot
}
\keyword{internal}
