% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{methods}
\name{setOptions,scAlignClass-method}
\alias{setOptions,scAlignClass-method}
\title{Method assigns projected data to S4 slots}
\usage{
\S4method{setOptions}{scAlignClass}(self, options)
}
\value{
scAlign class with options slot set
}
\description{
Method assigns projected data to S4 slots
}
\keyword{internal}
