% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{setOptions}
\alias{setOptions}
\title{Generic assigns projected data to S4 slots}
\usage{
setOptions(self, options)
}
\value{
scAlign class with options slot set
}
\description{
Generic assigns projected data to S4 slots
}
\keyword{internal}
