% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{methods}
\name{setProjected,scAlignClass-method}
\alias{setProjected,scAlignClass-method}
\title{Method assigns projected data to S4 slots}
\usage{
\S4method{setProjected}{scAlignClass}(self, projected.data, data.name,
  projection.name)
}
\value{
scAlign class with projection slot set
}
\description{
Method assigns projected data to S4 slots
}
\keyword{internal}
