% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\name{setProjected}
\alias{setProjected}
\title{Generic assigns projected data to S4 slots}
\usage{
setProjected(self, projected.data, data.name, projection.name)
}
\value{
scAlign class with projection slot set
}
\description{
Generic assigns projected data to S4 slots
}
\keyword{internal}
