% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlignClass.R
\docType{methods}
\name{show,scAlignClass-method}
\alias{show,scAlignClass-method}
\title{Method prints S4 class nicely}
\usage{
\S4method{show}{scAlignClass}(object)
}
\value{
Modified print statment for scAlign class objects
}
\description{
Method prints S4 class nicely
}
\keyword{internal}
