% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Smooth_by_chromosome}
\alias{Smooth_by_chromosome}
\title{Smoothing by chromosome using moving average}
\usage{
Smooth_by_chromosome(data, window_length = 101, gene_locs)
}
\arguments{
\item{data}{Matrix}

\item{window_length}{Integer}

\item{gene_locs}{Data.frame with columns: gene, chr, start, end}
}
\value{
Smoothed matrix
}
\description{
Smoothing by chromosome using moving average
}
