% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{smooth_helper}
\alias{smooth_helper}
\title{Helper for smoothing window (internal)}
\usage{
smooth_helper(obs_data, window_length)
}
\description{
Helper for smoothing window (internal)
}
\keyword{internal}
