% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{smooth_window}
\alias{smooth_window}
\title{Smoothing window for matrix}
\usage{
smooth_window(data, window_length)
}
\arguments{
\item{data}{Matrix}

\item{window_length}{Integer}
}
\value{
Smoothed matrix
}
\description{
Smoothing window for matrix
}
