% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdna_visualiztion.R
\name{SortCells_in_subclone}
\alias{SortCells_in_subclone}
\title{Sort cells within a subclone by correlation
This function sorts cells within a subclone based on a correlation vector.}
\usage{
SortCells_in_subclone(scdna_matrix, distance)
}
\arguments{
\item{scdna_matrix}{A matrix of CNV values (features x cells)}

\item{distance}{A numeric vector of correlation or distance values for sorting}
}
\value{
A matrix with columns (cells) sorted by distance
}
\description{
Sort cells within a subclone by correlation
This function sorts cells within a subclone based on a correlation vector.
}
