% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{subset_seurat_by_condition_or_phase}
\alias{subset_seurat_by_condition_or_phase}
\title{Subset Seurat object by condition or Phase}
\usage{
subset_seurat_by_condition_or_phase(seu, condition_value = "tumor")
}
\arguments{
\item{seu}{Seurat object}

\item{condition_value}{Character, e.g. 'tumor'}
}
\value{
Subsetted Seurat object
}
\description{
Subset Seurat object by condition or Phase
}
