% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Subtract_ref_mean_exp}
\alias{Subtract_ref_mean_exp}
\title{Subtract average reference expression}
\usage{
Subtract_ref_mean_exp(data)
}
\arguments{
\item{data}{Matrix}
}
\value{
Centered matrix
}
\description{
Subtract average reference expression
}
