% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCCA.R
\name{UpdateW}
\alias{UpdateW}
\title{Seurat V2 MultiCCA}
\usage{
UpdateW(mat.list, i, num.sets, ws, ws.final)
}
\arguments{
\item{mat.list}{List of matrix objects}

\item{i}{Optimization parameter}

\item{num.sets}{Number of sets}

\item{ws}{optim}

\item{ws.final}{optim}
}
\value{
CCA results
}
\description{
Seurat V2 MultiCCA
}
\keyword{internal}
