% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architectures.R
\name{encoder_large}
\alias{encoder_large}
\title{Network large}
\usage{
encoder_large(inputs, input_size = NULL, complexity = 3,
  emb_size = 32, l2_weight = 1e-04, dropout_rate = 0.3,
  is_training = TRUE, batch_norm = TRUE)
}
\value{
Neural network graph op
}
\description{
Defines network architecture for scAlign.
}
\keyword{internal}
