% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAlign_normalize_scrna_matrix.R
\name{Filter_genes_below_mean_exp_cutoff}
\alias{Filter_genes_below_mean_exp_cutoff}
\title{Filter genes below mean expression cutoff}
\usage{
Filter_genes_below_mean_exp_cutoff(mat, cutoff = 0.1)
}
\arguments{
\item{mat}{Expression matrix}

\item{cutoff}{Numeric, mean expression cutoff}
}
\value{
Filtered matrix
}
\description{
Filter genes below mean expression cutoff
}
