% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_seurat_processing.R
\name{find_and_plot_markers}
\alias{find_and_plot_markers}
\title{Find markers and plot heatmaps for merged Seurat object}
\usage{
find_and_plot_markers(aggr)
}
\arguments{
\item{aggr}{Processed Seurat object}
}
\description{
Find markers and plot heatmaps for merged Seurat object
}
