% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotTSNE}
\alias{PlotTSNE}
\title{Creates tsne plot}
\usage{
PlotTSNE(object, data.use, labels.use, cols = NULL, title = "",
  legend = "none", seed = 1234, ...)
}
\arguments{
\item{object}{scAlign class object with aligned data}

\item{data.use}{Specifies which alignment results to use.}

\item{labels.use}{Specifies "dataset" or "celltype" labeling from object meta.data.}

\item{cols}{Colours for plot}

\item{title}{ggplot title}

\item{legend}{Determines if legend should be drawn}

\item{seed}{Random seed for reproducability}

\item{...}{Additional arguments to Rtsne function}

\item{labels}{Object labels}
}
\value{
ggplot2 object
}
\description{
Creates tsne plot
}
\examples{

library(Seurat)
library(SingleCellExperiment)

## Input data, 1000 genes x 100 cells
data = matrix(sample.int(10000, 1000*100, TRUE), 1000, 100)
rownames(data) = paste0("gene", seq_len(1000))
colnames(data) = paste0("cell", seq_len(100))

age    = c(rep("young",50), rep("old",50))
labels = c(c(rep("type1",25), rep("type2",25)), c(rep("type1",25), rep("type2",25)))

ctrl.data = data[,which(age == "young")]
stim.data = data[,which(age == "old")]

ctrl = CreateSeuratObject(raw.data = ctrl.data, project = "MOUSE_AGE", min.cells = 0)
ctrl@meta.data$stim  = "YOUNG"
ctrl@meta.data$label = labels[which(age == "young")]
ctrl = ScaleData(ctrl, do.scale=TRUE, do.center=TRUE, scale.max=50, display.progress = TRUE)

stim = CreateSeuratObject(raw.data = stim.data, project = "MOUSE_AGE", min.cells = 0)
stim@meta.data$stim = "OLD"
stim@meta.data$label = labels[which(age == "old")]
stim = ScaleData(stim, do.scale=TRUE, do.center=TRUE, scale.max=50, display.progress = TRUE)

## Build the SCE object for input to scAlign using Seurat preprocessing and variable gene selection
ctrl.sce <- SingleCellExperiment(
              assays = list(
                counts = ctrl@raw.data,
                scale.data = ctrl@scale.data),
              colData = ctrl@meta.data)

stim.sce <- SingleCellExperiment(
              assays = list(
                counts = stim@raw.data,
                scale.data = stim@scale.data),
              colData = stim@meta.data)

## Build the scAlign class object and compute PCs
scAlignHSC = scAlignCreateObject(sce.objects = list("YOUNG"=ctrl.sce,
                                                    "OLD"=stim.sce),
                                 labels = list(ctrl.sce@colData@listData$label,
                                               stim.sce@colData@listData$label),
                                 pca.reduce = TRUE,
                                 pcs.compute = 50,
                                 cca.reduce = TRUE,
                                 ccs.compute = 15,
                                 project.name = "scAlign_Kowalcyzk_HSC")

## Run scAlign with high_var_genes
scAlignHSC = scAlign(scAlignHSC,
                    options=scAlignOptions(steps=1, log.every=1, early.stop=FALSE, architecture="large"),
                    encoder.data="scale.data",
                    supervised='none',
                    run.encoder=TRUE,
                    run.decoder=TRUE,
                    log.results=FALSE,
                    device="CPU")

## Plot alignment for 3 input types
example_plot = PlotTSNE(scAlignHSC, "ALIGNED-GENE", "labels", title="scAlign-Gene", perplexity=30)

}
